resource intarface
