+++ lib/kakine.rb
@@ -1,6 +1,6 @@
 require "kakine/version"
 require 'kakine/cli'
-require 'kakine/cli/operation'
+require 'kakine/operation'
 require 'kakine/adapter'
 require 'kakine/resource'
 require 'kakine/security_group'
