+++ lib/kakine/cli.rb
@@ -18,15 +18,6 @@ module Kakine
     desc 'apply', "apply local configuration into OpenStack"
     def apply
       filename = options[:filename] ? options[:filename] : "#{options[:tenant]}.yaml"
-      operation = Kakine::CLI::Operation.new
-
-      adapter = if options[:dryrun]
-        Kakine::Adapter::Mock.new
-      else
-        Kakine::Adapter::Real.new
-      end
-
-      operation.set_adapter(adapter)
 
       register_sg = Kakine::Resource.load_security_group_by_yaml(filename, options[:tenant])
       current = Kakine::Resource.get_current(options[:tenant])
